+++ DatatableJS/DatatableJS/Models/CaptionsModel.cs
@@ -1,8 +1,0 @@
-﻿namespace DatatableJS
-{
-    public class CaptionsModel
-    {
-        public string Top { get; set; }
-        public string Bottom { get; set; }
-    }
-}