+++ DatatableJS/DatatableJS/Models/FixedColumnsModel.cs
@@ -1,8 +1,0 @@
-﻿namespace DatatableJS
-{
-    public class FixedColumnsModel
-    {
-        public int LeftColumns { get; set; }
-        public int RightColumns { get; set; }
-    }
-}