+++ DatatableJS/DatatableJS/Models/LanguageModel.cs
@@ -1,7 +1,0 @@
-﻿namespace DatatableJS
-{
-    public class LanguageModel
-    {
-        public string URL { get; set; }
-    }
-}